Introduce context with CurrentUser & Params
The bones/web/context package wraps
github.com/gorilla/context and adds getters
and setters for type safe handling of values
that last the lifetime of the request.

CurrentUser is set by the Authenticate filter

Params is set by the Params filter, and contains
a pointer to a type that wraps *http.Request
with utility methods for fetching
path/query/form values

actions.FindEntityOr404 now takes the id as
a concrete int value instead of the name of the
parameter.