Add execdb package and example sql scripts
execdb is a tool for executing sql files/scripts.

If you don't give it a name of an sql script to
run, it will run all scripts listed in the
database.init file. This can be handy when you
need to recreate the database schema from a
clean slate (e.g during development, or first
deploy to production).

The tool does not enforce any naming conventions
for the sql script files. However, it is probably
a good idea to use names that indicate the
chronological order of changes.

It is also a good idea to include "undo" scripts
for every release, in case you need to roll back
the database to a previous version.